+++ framework/src/routes-compiler/src/main/twirl/play/routes/compiler/static/javaWrappers.scala.twirl
@@ -1,0 +1,28 @@
+@(sourceInfo: RoutesSourceInfo, packageName: String, controllers: Seq[String], reverseRefRouter: Boolean)
+// @@GENERATOR:play-routes-compiler
+// @@SOURCE:@sourceInfo.source
+// @@DATE:@sourceInfo.date
+
+@cb=@{"\u007D"}
+@ob=@{"\u007B"}
+
+package @packageName;
+
+public class routes @ob
+  @for(controller <- controllers) {
+  public static final @(packageName).Reverse@controller @controller = new @(packageName).Reverse@(controller)();}
+
+  public static class javascript @ob
+    @for(controller <- controllers) {
+    public static final @(packageName).javascript.Reverse@controller @controller = new @(packageName).javascript.Reverse@(controller)();}
+
+  @cb
+  @if(reverseRefRouter) {
+  public static class ref @ob
+    @for(controller <- controllers) {
+    public static final @(packageName).ref.Reverse@controller @controller = new @(packageName).ref.Reverse@(controller)();}
+
+  @cb
+  }
+
+@cb
