+++ framework/src/routes-compiler/src/main/twirl/play/routes/compiler/static/refReverseRouter.scala.twirl
@@ -1,0 +1,32 @@
+@(pkg: Option[String], routes: Seq[Route], namespaceReverseRouter: Boolean)
+
+@import play.routes.compiler._
+@import play.routes.compiler.templates._
+@tq=@{"\"\"\""}
+@cb=@{"\u007D"}
+@ob=@{"\u007B"}
+
+@for((packageName, routes) <- groupRoutesByPackage(routes)) {
+@markLines(routes: _*)
+package @(pkg.filter(_ => namespaceReverseRouter).fold(packageName)(_ + "." + packageName)).ref @ob
+@* This import statement is inserted mostly for the doc code samples, to ensure that controllers relative
+   to the namespace are in scope *@
+@for(p <- pkg if namespaceReverseRouter) {import @(p)._}
+@for((controller, routes) <- groupRoutesByController(routes)) {
+  @markLines(routes: _*)
+  class Reverse@controller.replace(".", "_") @ob
+  @for(((method, _), routes @ Seq(route, _*)) <- groupRoutesByMethod(routes)) {
+    @markLines(routes: _*)
+    def @(method)(@refReverseSignature(routes)): play.api.mvc.HandlerRef[_] = new play.api.mvc.HandlerRef(
+      @refCall(route),
+      HandlerDef(this.getClass.getClassLoader, "@pkg.getOrElse("")", "@(packageName).@(controller)",
+        "@method", Seq(@route.call.parameters.getOrElse(Nil).map("classOf[" + _.typeName + "]").mkString(", ")),
+        "@route.verb", @tq@route.comments.map(_.comment).mkString("\n")@tq, _prefix + @tq@route.path@tq
+      )
+    )
+  }
+  @cb
+}
+
+@cb
+}
