+++ framework/src/sbt-plugin/src/sbt-test/routes-compiler-plugin/source-mapping/Application.scala
@@ -1,0 +1,10 @@
+/*
+ * Copyright (C) 2009-2014 Typesafe Inc. <http://www.typesafe.com>
+ */
+package controllers
+
+import play.api.mvc._
+
+object Application extends Controller {
+  def index = Action(Ok)
+}
