+++ framework/src/sbt-plugin/src/sbt-test/routes-compiler-plugin/source-mapping/project/build.properties
@@ -1,0 +1,4 @@
+#
+# Copyright (C) 2009-2014 Typesafe Inc. <http://www.typesafe.com>
+#
+sbt.version=0.13.5
