+++ framework/src/sbt-plugin/src/sbt-test/routes-compiler-plugin/source-mapping/project/plugins.sbt
@@ -1,0 +1,4 @@
+//
+// Copyright (C) 2009-2014 Typesafe Inc. <http://www.typesafe.com>
+//
+addSbtPlugin("com.typesafe.play" % "sbt-plugin" % sys.props("project.version"))
