+++ framework/src/sbt-plugin/src/sbt-test/routes-compiler-plugin/source-mapping/test
@@ -1,0 +1,1 @@
+> all-problems-are-from routes 5
